feat(stat): use cputime for all timing related functions
